Fixed barrier logic and increased reset_bw_remaining to uint64

+++ src/cpu/testers/synchrotrace/synchro_trace.hh
@@ -480,7 +480,7 @@ class SynchroTraceReplayer : public MemObject
     uint8_t bw_remaining;
     // ADARSH this counts the number of times bw_remaining was reset
     // helps to debug how much data the function get/put from DM
-    uint8_t reset_bw_remaining;
+    uint64_t reset_bw_remaining;
 
     // bool variables indicating which phase is currently executing
     // for disaggr mem simulations; the entire sim is in_detailed mode
